+++ R/coerce.R
@@ -7,26 +7,13 @@ coerce <- function(x, type) {
 coerce_lgl <- function(x) coerce(x, "logical")
 coerce_int <- function(x) coerce(x, "integer")
 coerce_dbl <- function(x) coerce(x, "double")
-coerce_chr <- function(x) {
-  local_deprecation_user_env()
-  coerce(x, "character")
-}
-
-
-deprecate_to_char <- function(type) {
-  lifecycle::deprecate_warn(
-    "1.0.0",
-    I(paste0("Automatic coercion from ", type, " to character")),
-    I("an explicit call to `as.character()` within `map_chr()`"),
-    always = TRUE,
-    user_env = the$deprecation_user_env
-  )
-}
-
+coerce_chr <- function(x) coerce(x, "character")
 
 # Can rewrite after https://github.com/r-lib/rlang/issues/1643
-local_deprecation_user_env <- function(user_env = caller_env(2), frame = caller_env()) {
-
+local_deprecation_user_env <- function(
+  user_env = caller_env(2),
+  frame = caller_env()
+) {
   old <- the$deprecation_user_env
   the$deprecation_user_env <- user_env
   defer(the$deprecation_user_env <- old, frame)
@@ -37,4 +24,3 @@ defer <- function(expr, env = caller_env(), after = FALSE) {
   thunk <- as.call(list(function() expr))
   do.call(on.exit, list(thunk, TRUE, after), envir = env)
 }
-