+++ R/detect.R
@@ -2,10 +2,19 @@
 #'
 #' @inheritParams keep
 #' @inheritParams map
+#' @param .f A function, specified in one of the following ways:
+#'
+#'   * A named function, e.g. `mean`.
+#'   * An anonymous function, e.g. `\(x) x + 1` or `function(x) x + 1`.
+#'   * A formula, e.g. `~ .x + 1`. Use `.x` to refer to the first argument. No
+#'     longer recommended.
+#'   * A string, integer, or list, e.g. `"idx"`, `1`, or `list("idx", 1)` which
+#'     are shorthand for `\(x) pluck(x, "idx")`, `\(x) pluck(x, 1)`, and
+#'     `\(x) pluck(x, "idx", 1)` respectively. Optionally supply `.default` to
+#'     set a default value if the indexed element is `NULL` or does not exist.
 #' @param .dir If `"forward"`, the default, starts at the beginning of
 #'   the vector and move towards the end; if `"backward"`, starts at
 #'   the end of the vector and moves towards the beginning.
-#' @param .right `r lifecycle::badge("deprecated")` Please use `.dir` instead.
 #' @param .default The value returned when nothing is detected.
 #' @return `detect` the value of the first item that matches the
 #'  predicate; `detect_index` the position of the matching item.
@@ -24,8 +33,7 @@
 #' 3:10 |> detect_index(is_even, .dir = "backward")
 #'
 #'
-#' # Since `.f` is passed to as_mapper(), you can supply a
-#' # lambda-formula or a pluck object:
+#' # Since `.f` is passed to as_mapper(), you can supply a pluck object:
 #' x <- list(
 #'   list(1, foo = FALSE),
 #'   list(2, foo = TRUE),
@@ -41,11 +49,17 @@
 #'
 #' # If you need to find all positions, use map_lgl():
 #' which(map_lgl(x, "foo"))
-detect <- function(.x, .f, ..., .dir = c("forward", "backward"), .right = NULL, .default = NULL) {
+detect <- function(
+  .x,
+  .f,
+  ...,
+  .dir = c("forward", "backward"),
+  .default = NULL
+) {
   .f <- as_predicate(.f, ..., .mapper = TRUE)
   .dir <- arg_match0(.dir, c("forward", "backward"))
 
-  for (i in index(.x, .dir, .right, "detect")) {
+  for (i in index(.x, .dir, "detect")) {
     if (.f(.x[[i]], ...)) {
       return(.x[[i]])
     }
@@ -56,11 +70,11 @@ detect <- function(.x, .f, ..., .dir = c("forward", "backward"), .right = NULL, 
 
 #' @export
 #' @rdname detect
-detect_index <- function(.x, .f, ..., .dir = c("forward", "backward"), .right = NULL) {
+detect_index <- function(.x, .f, ..., .dir = c("forward", "backward")) {
   .f <- as_predicate(.f, ..., .mapper = TRUE)
   .dir <- arg_match0(.dir, c("forward", "backward"))
 
-  for (i in index(.x, .dir, .right, "detect_index")) {
+  for (i in index(.x, .dir, "detect_index")) {
     if (.f(.x[[i]], ...)) {
       return(i)
     }
@@ -70,18 +84,7 @@ detect_index <- function(.x, .f, ..., .dir = c("forward", "backward"), .right = 
 }
 
 
-
 index <- function(x, dir, right = NULL, fn) {
-  if (!is_null(right)) {
-    lifecycle::deprecate_warn(
-      when = "0.3.0",
-      what = paste0(fn, "(.right)"),
-      with = paste0(fn, "(.dir)"),
-      always = TRUE
-    )
-    dir <- if (right) "backward" else "forward"
-  }
-
   idx <- seq_along(x)
   if (dir == "backward") {
     idx <- rev(idx)