+++ R/imap.R
@@ -10,9 +10,14 @@
 #'   * A named function, e.g. `paste`.
 #'   * An anonymous function, e.g. `\(x, idx) x + idx` or
 #'     `function(x, idx) x + idx`.
-#'   * A formula, e.g. `~ .x + .y`. You must use `.x` to refer to the
-#'     current element and `.y` to refer to the current index. Only recommended
-#'     if you require backward compatibility with older versions of R.
+#'   * A formula, e.g. `~ .x + .y`. Use `.x` to refer to the current element and
+#'     `.y` to refer to the current index. No longer recommended.
+#'
+#'   `r lifecycle::badge("experimental")`
+#'
+#'   Wrap a function with [in_parallel()] to declare that it should be performed
+#'   in parallel. See [in_parallel()] for more details.
+#'   Use of `...` is not permitted in this context.
 #' @inheritParams map
 #' @return A vector the same length as `.x`.
 #' @export