+++ R/keep.R
@@ -15,9 +15,9 @@
 #'
 #'   * A named function, e.g. `is.character`.
 #'   * An anonymous function, e.g. `\(x) all(x < 0)` or `function(x) all(x < 0)`.
-#'   * A formula, e.g. `~ all(.x < 0)`. You must use `.x` to refer to the first
-#'     argument). Only recommended if you require backward compatibility with
-#'     older versions of R.
+#'   * A formula, e.g. `~ all(.x < 0)`. Use `.x` to refer to the first argument.
+#'     No longer recommended.
+#'
 #' @seealso [keep_at()]/[discard_at()] to keep/discard elements by name.
 #' @param ... Additional arguments passed on to `.p`.
 #' @export
@@ -26,7 +26,7 @@
 #'   map(sample, 5) |>
 #'   keep(function(x) mean(x) > 6)
 #'
-#' # Or use a formula
+#' # Or use shorthand form
 #' rep(10, 10) |>
 #'   map(sample, 5) |>
 #'   keep(\(x) mean(x) > 6)
@@ -63,8 +63,14 @@ compact <- function(.x, .p = identity) {
 
 #' Keep/discard elements based on their name/position
 #'
-#' @inheritParams map_at
+#' @description
+#' `keep_at()` and `discard_at()` are similar to `[` or `dplyr::select()`: they
+#' return the same type of data structure as the input, but only containing
+#' the requested elements. (If you're looking for a function similar to
+#' `[[` see [pluck()]/[chuck()]).
+#'
 #' @seealso [keep()]/[discard()] to keep/discard elements by value.
+#' @inheritParams map_at
 #' @export
 #' @examples
 #' x <- c(a = 1, b = 2, cat = 10, dog = 15, elephant = 5, e = 10)
@@ -72,8 +78,8 @@ compact <- function(.x, .p = identity) {
 #' x |> discard_at(letters)
 #'
 #' # Can also use a function
-#' x |> keep_at(~ nchar(.x) == 3)
-#' x |> discard_at(~ nchar(.x) == 3)
+#' x |> keep_at(\(x) nchar(x) == 3)
+#' x |> discard_at(\(x) nchar(x) == 3)
 keep_at <- function(x, at) {
   where <- where_at(x, at, user_env = caller_env())
   x[where]