+++ R/map2.R
@@ -10,10 +10,15 @@
 #'
 #'   * A named function.
 #'   * An anonymous function, e.g. `\(x, y) x + y` or `function(x, y) x + y`.
-#'   * A formula, e.g. `~ .x + .y`. You must use `.x` to refer to the current
-#'     element of `x` and `.y` to refer to the current element of `y`. Only
-#'     recommended if you require backward compatibility with older versions
-#'     of R.
+#'   * A formula, e.g. `~ .x + .y`. Use `.x` to refer to the current
+#'     element of `x` and `.y` to refer to the current element of `y`.
+#'     No longer recommended.
+#'
+#'   `r lifecycle::badge("experimental")`
+#'
+#'   Wrap a function with [in_parallel()] to declare that it should be performed
+#'   in parallel. See [in_parallel()] for more details.
+#'   Use of `...` is not permitted in this context.
 #' @inheritParams map
 #' @inherit map return
 #' @family map variants
@@ -54,25 +59,46 @@ map2_chr <- function(.x, .y, .f, ..., .progress = FALSE) {
   map2_("character", .x, .y, .f, ..., .progress = .progress)
 }
 
-map2_ <- function(.type,
-                  .x,
-                  .y,
-                  .f,
-                  ...,
-                  .progress = FALSE,
-                  .purrr_user_env = caller_env(2),
-                  .purrr_error_call = caller_env()) {
+map2_ <- function(
+  .type,
+  .x,
+  .y,
+  .f,
+  ...,
+  .progress = FALSE,
+  .purrr_user_env = caller_env(2),
+  .purrr_error_call = caller_env()
+) {
+  .progress <- as_progress(
+    .progress,
+    user_env = .purrr_user_env,
+    caller_env = .purrr_error_call
+  )
+
   .x <- vctrs_vec_compat(.x, .purrr_user_env)
   .y <- vctrs_vec_compat(.y, .purrr_user_env)
 
   n <- vec_size_common(.x = .x, .y = .y, .call = .purrr_error_call)
-  args <- vec_recycle_common(.x = .x, .y = .y, .size = n, .call = .purrr_error_call)
+  args <- vec_recycle_common(
+    .x = .x,
+    .y = .y,
+    .size = n,
+    .call = .purrr_error_call
+  )
   .x <- args$.x
   .y <- args$.y
 
   names <- vec_names(.x)
 
   .f <- as_mapper(.f, ...)
+
+  if (running_in_parallel(.f)) {
+    attributes(args) <- list(
+      class = "data.frame",
+      row.names = if (is.null(names)) .set_row_names(n) else names
+    )
+    return(mmap_(args, .f, .progress, .type, .purrr_error_call, ...))
+  }
 
   i <- 0L
   with_indexed_errors(