+++ R/pmap.R
@@ -20,8 +20,13 @@
 #'   * A named function.
 #'   * An anonymous function, e.g. `\(x, y, z) x + y / z` or
 #'     `function(x, y, z) x + y / z`
-#'   * A formula, e.g. `~ ..1 + ..2 / ..3`. This syntax is not recommended as
-#'     you can only refer to arguments by position.
+#'   * A formula, e.g. `~ ..1 + ..2 / ..3`. No longer recommended.
+#'
+#'   `r lifecycle::badge("experimental")`
+#'
+#'   Wrap a function with [in_parallel()] to declare that it should be performed
+#'   in parallel. See [in_parallel()] for more details.
+#'   Use of `...` is not permitted in this context.
 #' @inheritParams map
 #' @returns
 #' The output length is determined by the maximum length of all elements of `.l`.
@@ -117,18 +122,31 @@ pmap_chr <- function(.l, .f, ..., .progress = FALSE) {
   pmap_("character", .l, .f, ..., .progress = .progress)
 }
 
-pmap_ <- function(.type,
-                  .l,
-                  .f,
-                  ...,
-                  .progress = FALSE,
-                  .purrr_user_env = caller_env(2),
-                  .purrr_error_call = caller_env()) {
+pmap_ <- function(
+  .type,
+  .l,
+  .f,
+  ...,
+  .progress = FALSE,
+  .purrr_user_env = caller_env(2),
+  .purrr_error_call = caller_env()
+) {
+  .progress <- as_progress(
+    .progress,
+    user_env = .purrr_user_env,
+    caller_env = .purrr_error_call
+  )
+
   .l <- vctrs_list_compat(.l, error_call = .purrr_error_call)
   .l <- map(.l, vctrs_vec_compat)
 
   n <- vec_size_common(!!!.l, .arg = ".l", .call = .purrr_error_call)
-  .l <- vec_recycle_common(!!!.l, .size = n, .arg = ".l", .call = .purrr_error_call)
+  .l <- vec_recycle_common(
+    !!!.l,
+    .size = n,
+    .arg = ".l",
+    .call = .purrr_error_call
+  )
 
   if (length(.l) > 0L) {
     names <- vec_names(.l[[1L]])
@@ -138,6 +156,15 @@ pmap_ <- function(.type,
 
   .f <- as_mapper(.f, ...)
 
+  if (running_in_parallel(.f)) {
+    attributes(.l) <- list(
+      names = names(.l),
+      class = "data.frame",
+      row.names = if (is.null(names)) .set_row_names(n) else names
+    )
+    return(mmap_(.l, .f, .progress, .type, .purrr_error_call, ...))
+  }
+
   call_names <- names(.l)
   call_n <- length(.l)
 
@@ -146,7 +173,17 @@ pmap_ <- function(.type,
     i = i,
     names = names,
     error_call = .purrr_error_call,
-    call_with_cleanup(pmap_impl, environment(), .type, .progress, n, names, i, call_names, call_n)
+    call_with_cleanup(
+      pmap_impl,
+      environment(),
+      .type,
+      .progress,
+      n,
+      names,
+      i,
+      call_names,
+      call_n
+    )
   )
 }