+++ tests/testthat/test-rate.R
@@ -78,4 +78,3 @@ test_that("rate_sleep() checks that rate is still valid", {
   expect_snapshot(rate_sleep(rd), error = TRUE)
   expect_snapshot(rate_sleep(rd), error = TRUE)
 })
-