+++ NEWS.md
@@ -1,10 +1,40 @@
 # purrr (development version)
 
-* Added a test to assert that `list_transpose()` does not work on data frames
-  (@KimLopezGuell, #1141, #1149).
-* Added `imap_vec()` (#1084)
-* `list_transpose()` inspects all elements to determine the correct
-  template if it's not provided by the user  (#1128, @krlmlr).
+* Formatted strings for the progress bar could only access the global environment. `map()`, `map2()`, and `pmap()` have been updated to add  `caller = .purrr_user_env` to `.progress` by default, allowing formatted strings to access to the current/parent environment (@jcolt45, #1078).
+
+* `as_mapper.default()` optimized by removing special named argument handling for primitive functions (@mtcarsalot, #1088).
+
+* `list_flatten()` gains an `is_node` parameter taking a predicate function that determines whether an input element is a node or a leaf (@salim-b, #1179).
+
+* `in_parallel()` now accepts objects, including helper functions, supplied to `...` for all locally-defined functions (#1208).
+
+* `in_parallel()` now works in conjunction with string and list values supplied to the `.progress` argument of map functions (#1203).
+
+* All functions that were soft-deprecated in purrr 1.0.0 are now fully deprecated. They will be removed in a future release. This includes: `invoke_*()`, `lift_*()`, `cross*()`, `prepend()`, `splice()`, `rbernoulli()`, `rdunif()`, `when()`, `update_list()`, `*_raw()`, `vec_depth()`.
+
+* `map_chr()` no longer coereces from logical, integer, or double to strings.
+
+* All functions and arguments deprecated in purrr 0.3.0 have now been removed. This includes `%@%`, `accumulate_right()`, `at_depth()`, `cross_d()`, `cross_n()`, `reduce2_right()`, and `reduce_right()`.
+
+# purrr 1.1.0
+
+* purrr now requires R >= 4.1, so we can rely on the base pipe and lambda
+  syntax (#1177).
+
+* purrr gains `in_parallel()` to support parallel and distributed maps, powered 
+  by {mirai}. See `?in_parallel` for more details (@shikokuchuo, #1163, #1185).
+
+# purrr 1.0.4
+
+# purrr 1.0.3
+
+* Varies fixed to bring purrr back into compliance with R CMD check (@shikokuchuo, @jayhesselberth).
+
+* Added missing `imap_vec()` (#1084)
+
+* `list_transpose()` now asserts that it does not work on data frames
+  (@KimLopezGuell, #1141, #1149), and inspects all elements to determine
+  the correct template if not provided by the user  (#1128, @krlmlr).
 
 # purrr 1.0.2